+++ src/index.tsx
@@ -13,7 +13,7 @@ import Users from './pages/Users';
 import SignIn from './pages/SignIn';
 import SignUp from 'pages/SignUp';
 import ChatRoom from 'pages/ChatRoom';
-import GroupChatList from './pages/GroupChatList';
+import GroupChatList from './pages/GroupChatList/GroupChatList';
 import Modal from 'react-modal';
 
 Modal.setAppElement('#root');
