+++ src/pages/GroupChatList/HeaderLayout/GenerateChat/GenerateChat.tsx
@@ -1,9 +1,9 @@
-import SearchBar from 'components/HeaderLayout/SearchBar';
+import SearchBar from '../SearchBar';
 import {MdClose} from 'react-icons/md';
 import styled from 'styled-components';
 import {theme} from 'styles/Theme';
 import UserCells from './UserCells';
-import {User} from '../../types/chatroom.types';
+import {User} from '../../../../types/chatroom.types';
 import {useEffect, useState} from 'react';
 import {getUsers} from 'api/users';
 import {makeChatRoom} from 'api/myChatRoom';
@@ -19,19 +19,20 @@ type InputStates = [{name: 'chatName'; state: string}, {name: 'pickedUser'; stat
 
 const GenerateChat = (props: ModalProps) => {
   const [userData, setUserData] = useState<[User[], User[]]>([[], []]);
+  const [searchUserData, setSearchUserData] = useState('');
+
   const [chatName, setChatName] = useState<string>('');
 
   const [inputStates, setInputStates] = useState<InputStates>([
     {name: 'chatName', state: 'default'},
     {name: 'pickedUser', state: 'default'},
   ]);
-  const navigate = useNavigate();
 
+  const navigate = useNavigate();
   const myInfo = useRecoilValue(userInformation);
 
   useEffect(() => {
     // 사용자 불러오기
-
     getUsers().then(res => {
       const filtered = (res as User[]).slice().filter(val => val.id !== myInfo.id);
       setUserData([filtered, []]);
@@ -83,7 +84,7 @@ const GenerateChat = (props: ModalProps) => {
         <StyledMain>
           <StyledUnit>
             <StyledLabel>사용자 선택하기</StyledLabel>
-            <SearchBar content="사용자를 검색해보세요" height="40"></SearchBar>
+            <SearchBar content="사용자를 검색해보세요" height="40" onSearchName={setSearchUserData}></SearchBar>
             <UserCells
               height="312px"
               marginTop="8px"
@@ -91,6 +92,7 @@ const GenerateChat = (props: ModalProps) => {
               allocatedData={userData[0]}
               subData={userData[1]}
               onToggleUser={setUserData}
+              searchUserData={searchUserData}
             ></UserCells>
           </StyledUnit>
           <StyledUnit>
